requirements.txt
Pillow==9.3.0
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
shiboken6==6.4.1
